requirements.txt
Werkzeug==0.14.1
requests==2.12.4
Flask_Mail==0.9.1
Flask_WTF==0.14.2
Flask_Bootstrap==3.3.7.1
hashids==1.2.0
SQLAlchemy==1.3.0
Flask_Migrate==2.4.0
Flask_SQLAlchemy==2.3.2
Flask_Login==0.4.1
Flask_Moment==0.7.0
Flask==1.0.2
WTForms==2.2.1
alembic==1.0.8
PyJWT==1.7.1
